discord-bot/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.30
PyNaCl==1.5.0
